Fix form required attributes redundancy and keyword input description
Why these changes are being introduced:

DAS has informed us that it is bad practice to use both `required`
and `aria-required`. They also suggested that we add an
`aria-describedby` to the keyword input that links to the `p`
tag below the site title.

Relevant ticket(s):

* [GDT-327](https://mitlibraries.atlassian.net/browse/GDT-327)

How this addresses that need:

This removes the redundant `aria-required` attributes and makes
the suggested labeling change for the keyword input.

Side effects of this change:

The labeling piece only affects GDT, as non-GDT apps don't currently
use the `site_title` partial. This feels acceptable, because the
keyword input is more clearly labeled in other TIMDEX UI applications.