Refactor search form JS
Why these changes are being introduced:

During usability testing, stakeholders noted that the form can
enter a state in which they keyword field is incorrectly required.
(E.g., the geodistance panel is open, but the keyword input still
has the `required` attribute.)

In our attempts to replicate this bug, we have intermittently gotten
the form into a similar out-of-sync state. However, we have not
been able to replicate it consistently. Part of the problem is
the poor structure of the search form JS, which makes the script
difficult to debug.

Relevant ticket(s):

* [GDT-302](https://mitlibraries.atlassian.net/browse/GDT-302)
* [TIMX-281](https://mitlibraries.atlassian.net/browse/TIMX-281)

How this addresses that need:

This refactors the search form JS to be more well structured,
concise, and DRY, with the intention of resolving two tickets:

* GDT-302, the bug mentioned above. To resolve this, we delay the
execution of the functions that trigger the keyword input state
change, so as to avoid the syncing issues we found in testing.
* TIMX-281, a bug in non-GDT form state that prevents the visual
changes to icon and placeholder when the advanced search panel
is open.

Because we can't consistently replicate the bug that GDT-302
seeks to address, I can't say with certainty that this fixes the
problem. I've tried click-testing it, and certain behavior that
had previously gotten it out of sync no longer does. At minimum,
the refactor should improve the maintainability of our codebase
and make it easier to debug any future issues that may emerge.

Side effects of this change:

`var` is intentionally used instead of `const` or `let`. When
form submission fails with a redirect, Rails seems not to trigger
a full page reload. This causes the JS interpreter to throw an
error for modifying variables that have already been declare, unless
we declare them with `var`.

Since `var` is still valid JS, albeit not ideal, it's preferable
at this point to overriding Rails (likely Turbo) behavior on
form submit.