Fix iOS text wrap bug
Why these changes are being introduced:

SwiftUI/iOS has some unusual behavior for list markers. It adds
line height to them, such that when we increase the font size
for a marker, the line height is dramatically increased. This
is particularly visible when the text next to the marker wraps,
which is common for TIMDEX UI record titles.

Relevant ticket(s):

* [GDT-246](https://mitlibraries.atlassian.net/browse/GDT-246)

How this addresses that need:

This implements a CSS workaround. First, we reset the line height
to 0 on `.result`, enforcing no line height for markers. Then,
we add a div to wrap the rest of the result (`.result-content`),
with a rule that sets the line height to 1.2.

Side effects of this change:

I don't forsee any, but it's a pretty ugly workaround to accommodate
a minor bug in a weird OS.