Add models for Primo Search API integration
Why these changes are being introduced:

The first iteration of USE UI will include the
ability to toggle between Ex Libris CDI and
TIMDEX results.

Relevant ticket(s):

* [USE-30](https://mitlibraries.atlassian.net/browse/USE-30)

How this addresses that need:

This adds a Primo Search model that calls the
Primo Search API, a Normalize Primo Record model
parses the each result and returns a normalized
record, and a Normalize Primo Results model that
batches record normalization.

Side effects of this change:

* For better or worse, much of the normalization
code has been repurposed from Bento. This felt
acceptable as we do not anticipate this to be a
long-term solution. However, when we also
normalize TIMDEX records, we'll want to revisit
the normalized fields.
* While working on this, I noticed that most of
the TIMDEX normalization logic happens in the
views. It would be better to move this to
a normalization model, as we've done here with
Primo. I opened [USE-73](https://mitlibraries.atlassian.net/browse/USE-73)
to address this.
* It's not technically a side effect of this
changeset, but it's worth noting that we still
periodically see some GeoData test failures on
random test runs. I've documented this in [USE-72](https://mitlibraries.atlassian.net/browse/USE-72).
* We'll want to keep on eye on which methods
error in case more guard clases are needed. We
may also want to refactor to use more safe
navigation.