Connect UI application to TACOS
** Why are these changes being introduced:

The search UI needs to be connected with TACOS, in order to get insight
about interventions and allow term categorization to continue.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/use-66

** How does this address that need:

This adds a route, controller, and model for the integration with TACOS,
along with defining some new environment variables. The approach is
similar, but not identical, to the integration present in the Bento app.

** Document any side effects to this change:

One oddity I can't explain is why the "tacos" cassette doesn't have any
record of a call to TACOS - only the call to TIMDEX for the requested
search is present. The related test passes, so everything worked and the
UI element is present - but no interaction is present in the cassette
(and thus made by the application when tested).