Add error handling to Tacos model
** Why are these changes being introduced:

I realized that the Tacos model as originally proposed had no error
handling when I refreshed a results page without my local TACOS app
running. The result was a pretty ugly blob shown to the user.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/use-66

** How does this address that need:

This adds two catch statements to the call to TACOS - one for an HTTP
error response, and one for a JSON parsing error.

It also refactors the Tacos model to allow a separately-defined HTTP
object, which we craft in the test suite to always return those two
error conditions.

** Document any side effects to this change:

This is yet another pattern for error handling, not using VCR cassettes
or webmock, neither of which I could get working this afternoon. I'm
not happy about that, but this does work.