Render separate results partial for GeoData
Why these changes are being introduced:

As USE and GeoData continue to diverge, it makes
less sense for them to share view partials. This
became clear when the USE tabbed results view
broke GeoData full record view due to its
implementation of Turbo frames. To fix this
problem in the same results view would require an
even more complex map of forked logic than we
already had in that view.

Relevant ticket(s):

* [USE-31](https://mitlibraries.atlassian.net/browse/USE-31)

How this addresses that need:

This adds a separate `results_geo` partial that
is rendered when the GeoData feature flag is
active.

Side effects of this change:

* Forked views are something we've tried to avoid,
but in this case it felt like the least bad option.
* The `search_summary` partial has been renamed
to `search_summary_geo`, and `search_summary_use`
has been renamed to `search_summary`. This better
matches our naming conventions for forked views.
* An runelated change to `result_primo` tidies up
a link.