Removes bot_challenge_page gem
This was an interesting idea, but it wasn't triggering enough for to warrant the frustration the upgrade process caused when the gem entirely changed under the hood. I'm sure it made sense, but it was easier to remove it than update the configs to work on newer versions. We may bring it back eventually if it stabilizes or our needs warrant it.