Cache Primo queries
Why these changes are being introduced:

The Primo API performs poorly. We should do what
we can to mitigate this.

Relevant ticket(s):

* [USE-76](https://mitlibraries.atlassian.net/browse/USE-76)

How this addresses that need:

This adds caching to Primo queries.

Side effects of this change:

None, but it's worth noting that we experience
many unique searches, and this will only help
with latency if USE UI has seen the query before.