Refactor how TIMDEX records are nomalized
Why these changes are being introduced:

Prior to USE, TIMDEX records were largely normalized in the view layer.
While not ideal, this made some sense given that TIMDEX was the sole
source we were working with in the UI.

Now that we have introduced Primo results in USE, we should normalize
TIMDEX records similarly, such that the two share a similar data
structure.

Relevant ticket(s):

* [USE-73](https://mitlibraries.atlassian.net/browse/USE-73)

How this addresses that need:

This introduces NormalizeTimdexReuslts and NormalizeTimdexRecord models
to parallel the Primo models. The normalization models share a structure
to the extent that it is meaningful. Source-specific fields that will
be used in the application are indicated as such.

Side effects of this change:

* Modifications to various parts of the view layer were necessary to
retrofit this change.
* In some cases, we are only mapping the necessary data. (E.g., TIMDEX
links array only includes source link.) This is subject to change as
we learn more about UX requirements.
* We are still using separate result partials for TIMDEX and Primo,
until we have a better sense of how different the requirements are for
those sources.
* The Primo result partial has been minimized to more closely resemble
the TIMDEX partial. This is also subject to change based on UX
requirements.