Render titles with source links for TIMDEX records
Why these changes are being introduced:

In TIMDEX UI, titles linked to full record views. That feature is not
currently implemented in USE UI.

Relevant ticket(s):

* [USE-99](https://mitlibraries.atlassian.net/browse/USE-99)

How this addresses that need:

This updates TIMDEX title links to use the `source_link` field.

Side effects of this change:

When developing TIMDEX UI, we decided not to test any view logic. I've
moved the logic in this change to the Search Helper in order to test it,
but we may want to discuss whether to revisit integration testing in
USE UI.