Refactor search form for visual design changes
Why these changes are being introduced:

GeoData and USE UI currently share search form input and submit fields.
These fields should be isolated for UXWS to easily make tweaks to the
form for USE UI that won't affect GeoData.

Relevant ticket(s):

* [USE-127](https://mitlibraries.atlassian.net/browse/USE-127)

How this addresses that need:

This conditionally renders a submit button with class
`geo-search-submit` if GeoData is enabled. If not, the button div has
the class `basic-search-submit`.

This also adds a conditional `inline` class to the input field if
GeoData is not enabled. This class does not actually do anything yet,
but it will make it easier for UXWS update the CSS such that the
button renders inline with the input.

Side effects of this change:

The ID and classes associated with the form element have been renamed
from `basic-search-form` to `search-form`. This more accurately captures
the element they target, since the form is used both for USE UI (basic
search) and GeoData (advanced search).