Isolate search forms for visual design changes
Why these changes are being introduced:

GeoData and USE UI currently share search form input and submit fields.
These fields should be isolated for UXWS to easily make tweaks to the
form for USE UI that won't affect GeoData.

Relevant ticket(s):

* [USE-127](https://mitlibraries.atlassian.net/browse/USE-127)

How this addresses that need:

This creates separate form partials for USE and GeoData to faciliate
changes to the USE form without affecting GeoData. It also moves the
USE search form to the header, which is where UXWS' designs have it.

Side effects of this change:

It's possible we may want to move the GeoData form to the header as
well, but given the tight timeline for USE UI beta, it makes sense to
focus on the USE side of things for now.