Render FRBR links for Alma records only
Why these changes are being introduced:

FRBR/dedup links work for Alma records only. Even
though the frbrgroupid field sometimes shows up in
CDI records, if we construct a URL in the same way
as we do with Alma records, it results in a 404.

Relevant ticket(s):

* [USE-96](https://mitlibraries.atlassian.net/browse/USE-96)

How this addresses that need:

This adds a check for whether a given record is
an Alma record, by (somewhat unsatisfyingly)
confirming that its identifier begins with 'alma'.
Only records that pass this check get dedup URLs.

Side effects of this change:

Ideally, we would also have functional dedup URLs
for CDI records, but this feels like an adequate
solution for now.