Move tab filters to search form
Why these changes are being introduced:

UXWS would like the search tabs to display at the
bottom of the header, below the search form.

Relevant ticket(s):

* [USE-147](https://mitlibraries.atlassian.net/browse/USE-147)

How this addresses that need:

This moves the tabs to a partial, which is then rendered in the search
form if results are present. No additional conditional logic is
required, because the form itself is rendered only if the GeoData
feature is not enabled.

Side effects of this change:

None.