Show correct search form in GeoData full records
Why these changes are being introduced:

USE UI uses a different search form partial than
GeoData. When this change was made, we did not
update the form partial loaded in the full record
view.

Relevant ticket(s):

* [USE-157](https://mitlibraries.atlassian.net/browse/USE-157)

How this addresses that need:

This updates the `record_geo` partial to load the `form_geo` partial
when the GeoData feature is enabled.

Side effects of this change:

None.