Show Primo continuation partial only beyond page 1
Why these changes are being introduced:

A condition in the `load_primo_results` method in the Search Controller
renders the Primo continuation partial if results are empty. This is
necessary because of erratic behavior by the Primo search API, but it is
problematic because the condition also triggers if the search returns no
results.

Relevant ticket(s):

- [USE-152](https://mitlibraries.atlassian.net/browse/USE-152)

How this addresses that need:

This updates the problematic branch of the conditional so it also checks
if the current page is 2 or higher before rendering the Primo
continuation partial.

Side effects of this change:

None.