Add openssl to gemfile
Why these changes are being introduced:

The openssl gem is included in the Ruby standard library, but there is a
[bug](https://github.com/ruby/openssl/issues/949) in the included
version (`3.3.0`) that causes it to fail with OpenSSL 3.6.0.

Relevant ticket(s):

N/A.

How this addresses that need:

This includes the openssl gem in the Gemfile, thus installing the latest
version rather than what is provided by standard Ruby. This is
[recommended by openssl gem maintainers](https://github.com/ruby/openssl/issues/949#issuecomment-3381446772).

Side effects of this change:

This is something to remember for future perplexing bugs, because this
issue could happen with any gem included in the Ruby standard library.