Define new helper for tab links, with aria
** Why are these changes being introduced:

We want to add an `aria-current="page"` attribute to the currently
active tab link on a results page, so that assistive technologies can
parse which is the active tab. This requires a conditional link_to call,
as well as an update to the javascript which handles tab changes.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/use-180

** How does this address that need:

Because we can't (apparently) write a conditional inside an argument to
a link_to call, we have to maintain the conditional check elsewhere -
so to prevent the source_tabs partial from becoming a mess of multiple
conditionas, we define a new search helper function, link_to_tab. This
takes a string as an argument, and is capable of outputting both a link
to the active tab, as well as to an inactive tab.

We also update loading_spinner.js to be able to maintain the state of
the aria-current attribute.

** Document any side effects to this change:

The tests for this helper function _should_ be robust enough, but I
worry that they're a bit too clever. I didn't want to just write a
string comparison assertion, because that feels very fragile given how
much this partial has been changing lately. The assertions as written
should focus on the values being present or not, but still be robust to
other changes.

Also note that, for now, the source_tabs partial is still a bit bespoke
in that there are explicitly two calls to the various tabs. I'd
originally intended to define a list in the controller, but managing the
list of tabs in this way feels awkward and like it doesn't gain anything
since the tabs are entwined with so much of the rest of the application.
Abstracting the tab order and contents to a list doesn't save us from
other complexity.