Remove GeoData from active tab logic
Why these changes are being introduced:

With the implementation of the 'all' tab, which
includes a `valid_tab?` method, it's become
confusing as to how GeoData fits into the tab
system. We currently treat it as a tab, but it's
also treated as a feature.

Relevant ticket(s):

- [USE-176](https://mitlibraries.atlassian.net/browse/USE-176)
- [USE-177](https://mitlibraries.atlassian.net/browse/USE-177)

How this addresses that need:

This normalizes the GeoData feature flagging in
the controller, such that we only check for
`Feature.enabled?` when toggling GeoData, rather
than a combination of `active_tab` and the
`Feature` model.

Side effects of this change:

Remaining instances where 'GeoData/geodata' was
referred to as 'GDT/gdt' have been updated to use
the actual property name.