Fix bug that skips rendering single contributors
** Why are these changes being introduced:

Our rendering of contributors from TIMDEX apparently has a bug that
skips rendering anything if there is only one contributor on the record.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/use-214

** How does this address that need:

This adds an .each call to the .uniq filter that exists, which forces
Rails to actually process the single resulting record in the case of a
single contributor being present on the record.

** Document any side effects to this change:

Hopefully none.