Add pagination support to 'all' tab
Why these changes are being introduced:

We intentionally decided to implement pagination for the 'all' tab
separately to avoid excessively large tickets.

Relevant ticket(s):

- [USE-178](https://mitlibraries.atlassian.net/browse/USE-178)
- [USE-179](https://mitlibraries.atlassian.net/browse/USE-179)

How this addresses that need:

This updates the Search Controller and Analyzer Model to handle
pagination from multiple sources ('all' tab) in addition to single
sources.

Side effects of this change:

I found it very difficult to follow where and how the 'per-page' value
is set. I tried to improve the situation by moving this value to env.
I'm not sure it fully makes sense it, but I doubt it will without a
full refactor of the data flow architecture.