Add / remove spinner on pagination changes
** Why are these changes being introduced:

We recently refactored how the spinner is applied and removed when
changing between tabs, but neglected to be consistent when it comes to
pagination changes.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/use-201

** How does this address that need:

This adds the same add/remove calls for managing the spinner to the
pagination management branches of the javascript file. It elects to
maintain the separate pathways for these calls, rather than trying to
handle the spinner in a unified area of the event listeners.

** Document any side effects to this change:

The duplication of calls, like lines 46-50 being basically the same as
lines 26-30, may be seen as a code smell later. For now, though, I
prefer to keep them separate like this - both for parity between these
pathways and also symmetry with the .classList.add() lines at 61 and 75.

It seems safer to handle the spinner only in qualified contexts, rather
than risking it appearing or disappearing without that being intended.