Adjust normalized location in primo to use hash
We should move this logic to `holdings` as TIMDEX and Primo are currently normalizing different types
of information to a single field. TIMDEX has both `locations` and `holdings` information available whereas Primo only has `locations`. The information in TIMDEX `holdings` aligns well with what Primo `locations` is doing for us here.

Since we are not using TIMDEX locations or holdings in search results that use the normalizer, I have opted to resolve that under a separate ticket rather than fixing as part of this work. The data structure I am using here for Primo locations should align well with data available in TIMDEX holdings when we move both to USE normalized holdings.