Create integration with Libkey
** Why are these changes being introduced:

The application needs to look up DOI or PMID values using the Libkey
service in order to find the best possible fulfillment links for records
that come back from our various indices.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/use-232

** How does this address that need:

This defines a route, controller, model, and view template that will
receive a DOI or PMID, look them up in Libkey, and return the suggested
link.

Libkey's response object includes a "best integrator link", which we are
choosing here to accept uncritically.

There are two checks performed at various levels of this integration:
- The controller makes sure that Libkey is enabled, by making sure that
the proper env vars have been defined.
- The model makes sure that the proper parameters have been passed - we
require both a type (either "doi" or "pmid") and an identifier. The
format of the identifier is not specified. This value is used to create
the URL that is sent to Libkey, as part of the URL path.

The env vars created here are optional, because GeoData doesn't use
this integration.

** Document any side effects to this change:

Rubocop is flagging the lookup method as being too complex and too long.
I'm not sure the best way to resolve this, other than to split it into
"part A, setup" and "part B, communication" - but this feels like it
might be a distinction without a difference. Maybe a common "external
connection" model class could inform both the Tacos and Libkey models,
since they both use similar error handling patterns?

There is a third value in the Libkey response, "link type", which is
currently unused. I've flagged it as potentially being unneeded, but
wanted to make sure Dave / UX knows it exists.