Load libkey lookup results in UI
** Why are these changes being introduced:

Now that there's an integration with libkey via an internal path, the
search results page needs to consult that when relevant results are
received from our discovery APIs.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/use-232

** How does this address that need:

This adds fields for DOIs and PMIDs to the Primo record normalizer, on
the assumption that any given record has at most one of each of these.

If none is found, the field is left out of the normalized record.

If one is found that value is added to the normalized record.

If more than one is found, a message is sent to Sentry so we can look
further. This will help us understand how widespread the condition is.

(Any record might have _both_ a DOI and PMID, this is not problematic
and would not generate any Sentry messages).

When a record has one of these values, the view template that renders
the result calls the internal lookup route via a stimulus controller. If
the lookup results in an integration link (not all DOIs or PMIDs will
turn out this way), the link is added to the UI, after the other buttons
but before the availability text.

If a record has both a DOI and a PMID, only the DOI is used.

I've written tests for the normalizer model, to confirm that it handles
cases with zero, one, and mulitple of each field. The tests don't deal
with the Sentry integration, however.

** Document any side effects to this change:

This is the second place, in this ticket, that we are adding a direct
call to Sentry when certain data conditions appear. At some point if
this is a pattern we decide to keep using, a helper method might be
appropriate.