Remove cookies for active tab
Why these changes are being introduced:

UXWS has determined that the 'sticky tab' behavior is frustrating in
practice, and would prefer instead that the application identify the
current tab based on params.

Relevant ticket(s):

- [USE-228](https://mitlibraries.atlassian.net/browse/USE-228)

How this addresses that need:

This refactors the `set_active_tab` method to check for the presence of
`params[:tab]`. If that exists, it uses it as the active tab; if not, it
defaults to 'all'.

Side effects of this change:

None.