initial UI tweaks after discussion with UXWS
This does two things:
- Adds a prototypal skeleton loader element, with inline styles so that
there's a placeholder in the UI until we get the response from Libkey.
It is possible that this placeholder simply goes away, or gets replaced
with whatever we get back.

- Adds an inlined style to the resulting button that removes the border
from that element (at the moment, this button is given styles for the
"first child" pseudoclass, meaning the button gets a border that is not
desired).

Both of these tweaks can be polished further by Dave when he picks up
this work next.