Normalize fulfillment links
Why these changes are being introduced:

Fulfillment link logic exists in three separate
places. This is not necessary, as UXWS would like
the link text to always be 'View full record.'

Relevant ticket(s):

- [USE-252](https://mitlibraries.atlassian.net/browse/USE-252)

How this addresses that need:

This simplifies fulfillment link logic and
normalizes the text as requested.

Side effects of this change:

The nested conditional logic in the Primo result
partial is starting to get some code smell.
I find it unlikely that we'll want to render all
of the links long-term, though, so I think we
can accept this for now with the understanding
that it can probably be simplified later.