Move existing normalize method to a vocabulary
** Why are these changes being introduced:

We need a way to clean up messy format / content type values that come
back from Primo and TIMDEX. While ultimately we hope to fix data at the
source, it is also helpful to have a cleanup capability here.

We also want to standardize how these fields are treated across data
sources, because currently Primo, TIMDEX, and GeoData are all slightly
different.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/use-299

** How does this address that need:

We start by moving an existing "normalize_type" method out of the Primo
normalizer model and into a standalone class (a "format" vocabulary).

The new model includes additional entries mentioned in the ticket as
needing cleaned up.

There are no changes to the view layer here, because the "format" field
doesn't change shape - only where the work gets performed.

** Document any side effects to this change:

This also changes how the pnx.display.type field is handled, in that
each value is passed through the vocabulary individually, rather than
after having been joined together. It isn't clear to me whether Primo
ever actually returns multiple values, so this change may not impact
anything - but it bears mentioning, and will be needed when we get to
the TIMDEX side of things.